Working. Refactoring parser so its easy to extend.
The parser is currently a series of related functions depending on the record
type.  We need these to extract the same information and present it in the same
format, but this requires us to just make sure that each function is returning
the data in the same order. Easier option is to wrap up in a class and have a single
routine to do the printing. Plus I think this will make it easiser to add extra fields
later on which we know already that we want.