Extract code now more flexcible. Still migrating.
I've chosen not to use classes but instead have a bunch of 'switch'-lie statements
within code to extract fields of interest. We can easily add more fields and add record types.

I suspect I will move to a class based direction though as the current implementation looks
annoying to add new types of records. On the otherhand, there may not be many more to add.