Tested and passed all extraction commands.
Still not printing to flie but that should be easy now.

Ran into a problem when I started to process a native Uniprot file. Will add
another set of test cases.

Tests cases for different records are too long (>1000 lines). I'm going to
split this into multiple files.

Have added a script that can be run from command line to process inputs.
This will be critical to scaling this whole thing up.