Fixed no taxon_id, pseudo-gene and non-utf8 chars
Some more cornercases. In particular, we not ignore non-utf-8 characters,
thought it could be worthwhile running through all the genbank files and
checking for more of these. Could then report them to NCBI.

TODO: need to check that taxon being extracted is reasonable. Can do this by
parsing the node files that we've produced.